extract variant selector component
