+++ modules/ui/src/components/VariantSelector/index.ts
@@ -1,0 +1,1 @@
+export * from "./VariantSelector";
